+++ aries-site/src/data/structures/templates/index.js
@@ -1,0 +1,2 @@
+export * from './content-layouts';
+export * from './templates';
