--- aries-site/src/pages/templates/content-layout.mdx
@@ -1,3 +1,0 @@
-import { ComingSoon } from '../../components';
-
-<ComingSoon />
